Adds debounce to utils and adds barrel file
